backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
uvloop==0.19.0
websockets==13.0
opencv-python==4.10.0.84
ultralytics==8.3.0
torch
Pillow==10.4.0
google-generativeai==0.8.0
python-dotenv==1.0.1
aiosqlite==0.20.0
